Problem: Inconsistent key location across static clients on multiple web front-ends.
Possible Solution: Setting the IsVbucket property to true on MembasePool::InitVBucket

I might be missing something here, but since we're explicitly setting the Locator to be a VBucketNodeLocator, I would think we should set the IsVbucket property to true. Without this setting, the Ketama locator will be used when the node comes back online.  It seems possible that some of our static clients are seeing a node failure while others are not.  This should be OK as long as the node location process is the same accross all clients.  The V-Bucket system is well suited to this, but if some nodes are using V-Bucket and others are using Ketama, node location will not be consistent.  This change should resolve the issue.